requirements.txt
esphome==2025.8.0
Flask==3.1.2
Flask-Alembic==3.1.1
Flask-SQLAlchemy==3.1.1
PyYAML==6.0.2
wheel==0.45.1
